+++ BetterCombat/Data/RaceData.cs
@@ -1,0 +1,33 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace BetterCombat.Data
+{
+    internal class RaceData
+    {
+
+        #region BlueprintIds
+
+        internal static string aasimarRaceId            = "b7f02ba92b363064fb873963bec275ee";
+        internal static string duergarRaceId            = "cd40ff5a556bcf3419bf7479616cd2ad";
+        internal static string duergarTyrantRaceId      = "315a604df7e8cb34098ed2a3b7e3d068";
+        internal static string goblinRaceId             = "9d168ca7100e9314385ce66852385451";
+        internal static string dwarfRaceId              = "c4faf439f0e70bd40b5e36ee80d06be7";
+        internal static string elfRaceId                = "25a5878d125338244896ebd3238226c8";
+        internal static string gnomeRaceId              = "ef35a22c9a27da345a4528f0d5889157";
+        internal static string halfElfRaceId            = "b3646842ffbd01643ab4dac7479b20b0";
+        internal static string halflingRaceId           = "b0c3ef2729c498f47970bb50fa1acd30";
+        internal static string halfOrcRaceId            = "1dc20e195581a804890ddc74218bfd8e";
+        internal static string humanRaceId              = "0a5d473ead98b0646b94495af250fdc4";
+        internal static string tieflingRaceId           = "5c4e42124dc2b4647af6e36cf2590500";
+
+        #endregion
+
+        #region Localization
+        #endregion
+
+    }
+}