+++ BetterCombat/Data/SoftCoverData.cs
@@ -1,0 +1,31 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace BetterCombat.Data
+{
+    internal static class SoftCoverData
+    {
+        #region BlueprintIds
+
+        internal static string ImprovedPreciseShotFeatId = "46f970a6b9b5d2346b10892673fe6e74";
+
+        #endregion
+
+        #region Localization
+
+        internal static string SoftCoverNameKey = "8cd68cd5-3094-4cf0-8895-5187f36dbe2b";
+        internal static string SoftCoverName = "Soft Cover";
+
+        internal static string SoftCoverPartialNameKey = "04b56844-104e-4ad0-a64f-cd40646bfc28";
+        internal static string SoftCoverPartialName = "Soft Cover (Partial)";
+
+
+        internal static string ImprovedPreciseShotDescriptionId = "0c24e718-b23b-434c-987d-ec4a0e13acd6";
+        internal static string ImprovedPreciseShotDescription = "Your ranged attacks ignore the AC bonus granted to targets by anything less than total cover, and the miss chance granted to targets by anything less than total concealment.\nTotal cover and total concealment provide their normal benefits against your ranged attacks.";
+
+        #endregion
+    }
+}