+++ BetterCombat/CHANGELOG.md
@@ -7,5 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [1.0.2]
 
 ### Fixed
-- Fixed: Certain free combat maneuvers (Aspect of the Wolf, Piledriver) still provoke attacks of opportunity.
-- Fixed: New text (currently Soft Cover and the Improved Combat Maneuver feat names) will be added to the localization pack even if the locale is not English.
+- Certain free combat maneuvers (Aspect of the Wolf, Piledriver) still provoked attacks of opportunity.
+- New text (currently Soft Cover and the Improved Combat Maneuver feat names) will be added to the localization pack even if the locale is not English.
+
+## [1.1.0]
+
+### Added
+- Feat: [Low Profile](https://www.d20pfsrd.com/feats/combat-feats/low-profile-combat/)
+### Fixed
+- Updated Improved Precise Shot description to include cover.